refactor: combine StartedGame and GameContainer
